GODRIVER-2737 ConnectionId returned in heartbeats may be int64

+++ mongo/integration/mtest/opmsg_deployment.go
@@ -91,8 +91,8 @@ func (*connection) ID() string {
 }
 
 // ServerConnectionID returns a fixed identifier for the server connection.
-func (*connection) ServerConnectionID() *int32 {
-	serverConnectionID := int32(42)
+func (*connection) ServerConnectionID() *int64 {
+	serverConnectionID := int64(42)
 	return &serverConnectionID
 }
